Add endpoint documentation
Documentation for operations in our API, in adhoc format. This files
will later be used to generate a final documentation file. The
documentation for each endpoint is achieved on a test method, which we
use to make a basic test to the endpoint, and to specify documentation
on the endpoint.